Refactored definitions to own file. Removed Frends.Tasks.Atributes. Updated documentation and version number.
