+++ Frends.Community.WebForm.Tests/packages.config
@@ -1,4 +1,5 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
   <package id="NUnit" version="3.5.0" targetFramework="net452" />
+  <package id="NUnit3TestAdapter" version="3.10.0" targetFramework="net45" />
 </packages>
