+++ Frends.Community.WebForm.nuspec
@@ -2,7 +2,7 @@
 <package xmlns="http://schemas.microsoft.com/packaging/2011/10/nuspec.xsd">
   <metadata>
     <id>Frends.Community.WebForm</id>
-    <version>0.0.1</version>
+    <version>1.1.0</version>
     <title>FRENDS community web form task</title>
     <authors>HiQ Finland</authors>
     <owners />
@@ -17,5 +17,7 @@
   </metadata>
   <files>
     <file src="Frends.Community.WebForm\bin\Release\Frends.Community.WebForm.dll" target="lib\net452\Frends.Community.WebForm.dll" />
+    <file src="Frends.Community.WebForm\bin\Release\Frends.Community.WebForm.xml" target="lib\net452\Frends.Community.WebForm.xml" />
+
   </files>
 </package>
