+++ Frends.Community.WebForm/Frends.Community.WebForm.csproj
@@ -46,6 +46,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Definitions.cs" />
     <Compile Include="SubmitForm.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
