+++ Frends.Community.WebForm/SubmitForm.cs
@@ -3,7 +3,6 @@ using System.Net;
 using System.ComponentModel;
 using System.IO;
 using System.Net.Http;
-using Frends.Tasks.Attributes;
 
 namespace Frends.Community.WebForm
 {    /// <summary>
@@ -12,90 +11,10 @@ namespace Frends.Community.WebForm
     public static class SubmitForm
     {
         /// <summary>
-        /// Form parameters
-        /// </summary>
-        public class FormParameters
-        {
-            /// <summary>
-            /// Input name
-            /// </summary>
-            public string InputName { get; set; }
-            /// <summary>
-            /// Value
-            /// </summary>
-            public string Value { get; set; }
-        }
-
-        /// <summary>
-        /// Form file parameters if files are sent
-        /// </summary>
-        public class FormFileParameters
-        {
-            /// <summary>
-            /// Input name
-            /// </summary>
-            public string InputName { get; set; }
-            /// <summary>
-            /// File location
-            /// </summary>
-            public string FileLocation { get; set; }
-        }
-
-        /// <summary>
-        /// Form parameters
-        /// </summary>
-        public class Parameters
-        {
-            /// <summary>
-            /// List of form input parameters
-            /// </summary>
-            public FormParameters[] FormParameters { get; set; }
-            /// <summary>
-            /// List of form file input parameters
-            /// </summary>
-            public FormFileParameters[] FileParameteres { get; set; }
-
-        }
-
-        /// <summary>
-        /// Options for the call
-        /// </summary>
-        public class Options
-        {
-
-            /// <summary>
-            /// Web service address
-            /// </summary>
-            public string Address { get; set; }
-
-            /// <summary>
-            /// Username
-            /// </summary>
-            public string Username { get; set; }
-
-            /// <summary>
-            /// Password
-            /// </summary>
-            [PasswordPropertyText]
-            public string Password { get; set; }
-        }
-
-        /// <summary>
-        /// Return object
-        /// </summary>
-        public class Output
-        {
-            /// <summary>
-            /// Request result
-            /// </summary>
-            public HttpResponseMessage Result { get; set; }
-        }
-
-        /// <summary>
-        /// Send form data.
+        /// Fill and send data to web forms.
         /// </summary>
         /// <returns>Object {string FilePath }  </returns>
-        public static Output SendForm([CustomDisplay(DisplayOption.Tab)]Parameters parameters, [CustomDisplay(DisplayOption.Tab)]Options options)
+        public static Output SendForm(Parameters parameters, Options options)
         {
             HttpResponseMessage result = null;
             var handler = new HttpClientHandler();
