+++ README.md
@@ -1,26 +1,16 @@
 **[Table of Contents](http://tableofcontent.eu)**
 - [Frends.Community.WebForm](#frendscommunitywebsubmitform)
-  - [Contributing](#contributing)
   - [Documentation](#documentation)
     - [Parameters](#parameters)
     - [Options](#options)
     - [Result](#result)
+  - [Contributing](#contributing)
+  - [Change Log](#change-log)
   - [License](#license)
 
 
 # Frends.Community.WebForm
 FRENDS Task that posts web form.
-
-## Contributing
-When contributing to this repository, please first discuss the change you wish to make via issue, email, or any other method with the owners of this repository before making a change.
-
-1. Fork the repo on GitHub
-2. Clone the project to your own machine
-3. Commit changes to your own branch
-4. Push your work back up to your fork
-5. Submit a Pull request so that we can review your changes
-
-NOTE: Be sure to merge the latest from "upstream" before making a pull request!
 
 ## Documentation
 
@@ -45,6 +35,23 @@ NOTE: Be sure to merge the latest from "upstream" before making a pull request!
 |-----------------|----------|----------------------------------|
 | FilePath        | HttpResponseMessage   | webrequest result|
 
+## Contributing
+When contributing to this repository, please first discuss the change you wish to make via issue, email, or any other method with the owners of this repository before making a change.
+
+1. Fork the repo on GitHub
+2. Clone the project to your own machine
+3. Commit changes to your own branch
+4. Push your work back up to your fork
+5. Submit a Pull request so that we can review your changes
+
+NOTE: Be sure to merge the latest from "upstream" before making a pull request!
+
+## Change Log
+
+| Version | Changes |
+| ----- | ----- |
+| 1.0.0 | Initial version of web form task |
+| 1.1.0 | Frends.Tasks.Attributes library is no longer used. Updated documentation. |
 
 
 ## License
